Move 'src/domain/page' to 'src/app/page'
